Update AdvisorOptionHeaders.php
fixed small syntax thing

This version has error checking and messages
To do: make AdvisorCheckStudentAppointments.php and all the functionalities. 
AdvisorCheckStudentAppointments will check the appointments that the student has, to do similar error checking as in ValidateStudentSignin. But, the advisor will have the power to override these constraints (I need to talk to you guyss)